Write tests for instrumentation and fix a few bugs found along the way
Fix bug found for async execution mode when worker didn't correctly finish
shutting down.

Also, always instrument retry_all and release_many_blocked even if none is
retried/released, so we have that information as well, with the number of jobs
that actually were acted on.

+++ app/models/solid_queue/execution/dispatching.rb
@@ -9,8 +9,11 @@ module SolidQueue
         def dispatch_jobs(job_ids)
           jobs = Job.where(id: job_ids)
 
-          Job.dispatch_all(jobs).map(&:id).tap do |dispatched_job_ids|
-            where(job_id: dispatched_job_ids).order(:job_id).delete_all
+          Job.dispatch_all(jobs).map(&:id).then do |dispatched_job_ids|
+            if dispatched_job_ids.none? then 0
+            else
+              where(job_id: dispatched_job_ids).order(:job_id).delete_all
+            end
           end
         end
       end
